fix: change do_faccessat return type to long to match declaration in internal.h

+++ fs/open.c
@@ -359,7 +359,7 @@ extern int ksu_handle_faccessat(int *dfd, const char __user **filename_user,
  * We do this by temporarily clearing all FS-related capabilities and
  * switching the fsuid/fsgid around to the real ones.
  */
-int do_faccessat(int dfd, const char __user *filename, int mode)
+long do_faccessat(int dfd, const char __user *filename, int mode)
 {
 	const struct cred *old_cred;
 	struct cred *override_cred;
